+++ LykkePay/Models/OrderModels.cs
@@ -7,7 +7,7 @@ namespace LykkePay.Models
     public class OrderRequestModel
     {
         public string currency { get; set; }
-        public double amount { get; set; }
+        public decimal amount { get; set; }
         public string exchangeCurrency { get; set; }
         public string successURL { get; set; }
         public string errorURL { get; set; }
@@ -22,9 +22,9 @@ namespace LykkePay.Models
         public string address { get; set; }
         public string orderId { get; set; }
         public string currency { get; set; }
-        public double amount { get; set; }
-        public double recommendedFee { get; set; }
-        public double totalAmount { get; set; }
-        public double exchangeRate { get; set; }
+        public decimal amount { get; set; }
+        public decimal recommendedFee { get; set; }
+        public decimal totalAmount { get; set; }
+        public decimal exchangeRate { get; set; }
     }
 }
