+++ LykkePay/Models/PostAssetsPairRatesModel.cs
@@ -8,8 +8,8 @@ namespace LykkePay.Models
     public class PostAssetsPairRatesModel
     {
         public string assetPair { get; set; }
-        public double ask { get; set; }
-        public double bid { get; set; }
+        public decimal ask { get; set; }
+        public decimal bid { get; set; }
         public int accuracy { get; set; }
         [JsonProperty("Lykke-Merchant-Session-Id")]
         public string LykkeMerchantSessionId { get; set; }
